Add HashSet implementation
